// src/types/typechain-types/InfoContract.ts
export interface InfoContractInterface extends Interface {
	getFunction(
		nameOrSignature: 'setInfo' | 'sayHi' | 'getInfo'
	): FunctionFragment;

	getEvent(nameOrSignatureOrTopic: 'Instructor'): EventFragment;

	encodeFunctionData(
		functionFragment: 'setInfo',
		values: [string, BigNumberish]
	): string;
	encodeFunctionData(functionFragment: 'sayHi', values?: undefined): string;
	encodeFunctionData(functionFragment: 'getInfo', values?: undefined): string;

	decodeFunctionResult(functionFragment: 'setInfo', data: BytesLike): Result;
	decodeFunctionResult(functionFragment: 'sayHi', data: BytesLike): Result;
	decodeFunctionResult(functionFragment: 'getInfo', data: BytesLike): Result;
}

export namespace InstructorEvent {
	export type InputTuple = [name: string, age: BigNumberish];
	export type OutputTuple = [name: string, age: bigint];
	export interface OutputObject {
		name: string;
		age: bigint;
	}
	export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
	export type Filter = TypedDeferredTopicFilter<Event>;
	export type Log = TypedEventLog<Event>;
	export type LogDescription = TypedLogDescription<Event>;
}

export interface InfoContract extends BaseContract {
	connect(runner?: ContractRunner | null): InfoContract;
	waitForDeployment(): Promise<this>;

	interface: InfoContractInterface;

	queryFilter<TCEvent extends TypedContractEvent>(
		event: TCEvent,
		fromBlockOrBlockhash?: string | number | undefined,
		toBlock?: string | number | undefined
	): Promise<Array<TypedEventLog<TCEvent>>>;
	queryFilter<TCEvent extends TypedContractEvent>(
		filter: TypedDeferredTopicFilter<TCEvent>,
		fromBlockOrBlockhash?: string | number | undefined,
		toBlock?: string | number | undefined
	): Promise<Array<TypedEventLog<TCEvent>>>;

	on<TCEvent extends TypedContractEvent>(
		event: TCEvent,
		listener: TypedListener<TCEvent>
	): Promise<this>;
	on<TCEvent extends TypedContractEvent>(
		filter: TypedDeferredTopicFilter<TCEvent>,
		listener: TypedListener<TCEvent>
	): Promise<this>;

	once<TCEvent extends TypedContractEvent>(
		event: TCEvent,
		listener: TypedListener<TCEvent>
	): Promise<this>;
	once<TCEvent extends TypedContractEvent>(
		filter: TypedDeferredTopicFilter<TCEvent>,
		listener: TypedListener<TCEvent>
	): Promise<this>;

	listeners<TCEvent extends TypedContractEvent>(
		event: TCEvent
	): Promise<Array<TypedListener<TCEvent>>>;
	listeners(eventName?: string): Promise<Array<Listener>>;
	removeAllListeners<TCEvent extends TypedContractEvent>(
		event?: TCEvent
	): Promise<this>;

	setInfo: TypedContractMethod<
		[_name: string, _age: BigNumberish],
		[void],
		'nonpayable'
	>;

	sayHi: TypedContractMethod<[], [string], 'view'>;

	getInfo: TypedContractMethod<[], [[string, bigint]], 'view'>;

	getFunction<T extends ContractMethod = ContractMethod>(
		key: string | FunctionFragment
	): T;

	getFunction(
		nameOrSignature: 'setInfo'
	): TypedContractMethod<
		[_name: string, _age: BigNumberish],
		[void],
		'nonpayable'
	>;
	getFunction(
		nameOrSignature: 'sayHi'
	): TypedContractMethod<[], [string], 'view'>;
	getFunction(
		nameOrSignature: 'getInfo'
	): TypedContractMethod<[], [[string, bigint]], 'view'>;

	getEvent(
		key: 'Instructor'
	): TypedContractEvent<
		InstructorEvent.InputTuple,
		InstructorEvent.OutputTuple,
		InstructorEvent.OutputObject
	>;

	filters: {
		'Instructor(string,uint256)': TypedContractEvent<
			InstructorEvent.InputTuple,
			InstructorEvent.OutputTuple,
			InstructorEvent.OutputObject
		>;
		Instructor: TypedContractEvent<
			InstructorEvent.InputTuple,
			InstructorEvent.OutputTuple,
			InstructorEvent.OutputObject
		>;
	};
}